\name{skewness.test}
\alias{skewness.test}

\title{
Chi-square test skewness in the standardised residuals
}
\description{
Chi-square test skewness in the standardised residuals
}
\usage{
skewness.test(x)
}
\arguments{
  \item{x}{
numeric vector
}
}
\details{
}
\value{
Returns a list with the following elements:
  \item{statistic}{the test statistic}
  \item{p.value}{numeric between 0 and 1, the p-value of the test under the
  null of no-skewness}
}
\references{
}
\author{
See Sucarrat and Escribano (2011)
}
\note{
}
\seealso{
}
\examples{
}
\keyword{}
